improve agent error to be module-wide
Error submodules are preferred for error types, and it is preferable to
create multiple error classifications rather than bubbling all errors up
to a single error type. Therefore, improve agent error to do away with
crate-wide error modules.
