+++ agent/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod agentclient;
 pub mod apiclient;
-pub mod error;
